% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gapminder data.}
\format{
A data frame with 1704 rows and 6 variables:
\describe{
\item{country}{factor with 142 levels}
\item{continent}{factor with 5 levels}
\item{year}{ranges from 1952 to 2007 in increments of 5 years}
\item{lifeExp}{life expectancy at birth, in years}
\item{pop}{population}
\item{gdpPercap}{GDP per capita (US$, inflation-adjusted)}
}
}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
gapminder
}
\description{
Excerpt of the Gapminder data on life expectancy, GDP per capita, and
population by country.
}
\keyword{datasets}
